models now handle authentication
